added stye to the form and change the fields to admit just numher in salary range

+++ data_acid/blog/models.py
@@ -15,5 +15,5 @@ class Articles(models.Model):
 
 class LeadsForm(models.Model):
     profession = models.CharField(max_length=100)
-    salary_range = models.IntegerField()
+    salary_range = models.PositiveBigIntegerField()
     email = models.EmailField()
